+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 v0.2.0 (TBC)
 ............
 * allow annotation only fields to support mypy
+* add pretty ``to_string(pretty=True)`` method for models
 
 v0.1.0 (2017-06-03)
 ...................